feat:add reset password

+++ src/controllers/ProjectController.ts
@@ -11,53 +11,65 @@ export class ProjectController {
       res.status(500).send("Server error");
     }
   };
+
   static getProjectById = async (req: Request, res: Response) => {
     const { id } = req.params;
     try {
-      const project = await Project.findById(id);
+      const project = await Project.findById(id).populate("tasks");
       if (!project) {
-        const error = new Error(`Project not found`);
+        const error = new Error("Project not found");
         return res.status(404).json({ error: error.message });
       }
       res.json(project);
     } catch (error) {
       console.log(error);
+      res.status(500).send("Server error");
     }
   };
+
   static updatedProject = async (req: Request, res: Response) => {
     const { id } = req.params;
     try {
-      const project = await Project.findByIdAndUpdate(id, req.body);
+      const project = await Project.findByIdAndUpdate(id, req.body, {
+        new: true,
+      }).populate("tasks");
       if (!project) {
-        const error = new Error(`Project not found`);
+        const error = new Error("Project not found");
         return res.status(404).json({ error: error.message });
       }
-      await project.save();
       res.send("Project updated successfully");
     } catch (error) {
       console.log(error);
+      res.status(500).send("Server error");
     }
   };
+
   static deleteProject = async (req: Request, res: Response) => {
     const { id } = req.params;
     try {
       const project = await Project.findById(id);
+      if (!project) {
+        const error = new Error("Project not found");
+        return res.status(404).json({ error: error.message });
+      }
       await project.deleteOne();
       res.send("Project excluded");
     } catch (error) {
       console.log(error);
+      res.status(500).send("Server error");
     }
   };
+
   static createProject = async (req: Request, res: Response) => {
     const project = new Project(req.body);
     console.log(project);
 
     try {
       await project.save();
-
-      res.send("Project Created Success");
+      res.send("Project Created Successfully");
     } catch (error) {
       console.log(error);
+      res.status(500).send("Server error");
     }
   };
 }
